+++ packages/wiz-ui-next/src/components/base/inputs/search-input/types.ts
@@ -1,10 +1,12 @@
+export type CheckboxOption = string | number | string[]; // WizCheckBoxNew
+
 export type Tag = {
   label: string;
 };
 
-export type SearchInputOption = {
+export type SearchInputOption<T extends CheckboxOption = number> = {
   label: string;
-  value: number;
-  children?: SearchInputOption[];
+  value: T;
+  children?: SearchInputOption<T>[];
   tag?: Tag;
 };