+++ packages/wiz-ui-next/src/components/custom/chat/card/chat-card.stories.ts
@@ -1,4 +1,4 @@
-import { StoryFn, Meta } from "@storybook/vue3";
+import { Meta, StoryFn } from "@storybook/vue3";
 import { ref } from "vue";
 
 import { WizBox } from "@/components";