+++ packages/wiz-ui-react/src/components/base/inputs/search-input/components/search-input.tsx
@@ -193,38 +193,35 @@ const SearchInput = <T extends CheckboxOption>({
                 name={name}
                 disabled={disabled}
                 onChange={(e) => {
-                  setIsPopupOpen(true);
                   setFilteringText(e.target.value);
                 }}
                 onFocus={() => setIsFocused(true)}
                 onBlur={() => setIsFocused(false)}
-                onClick={() => setIsPopupOpen(!isPopupOpen)}
+                onClick={() => setIsPopupOpen(true)}
                 autoComplete="off"
               />
             </WizHStack>
           </WizHStack>
         </div>
       </div>
-      {filteredOptions.length > 0 && !disabled && (
-        <WizPopup
-          anchorElement={inputRef}
-          isOpen={isPopupOpen}
-          onClose={() => setIsPopupOpen(false)}
-          isDirectionFixed={isDirectionFixed}
-        >
-          <WizHStack nowrap>
-            <SearchPopupPanel
-              options={filteredOptions}
-              closePopup={() => setIsPopupOpen(false)}
-              values={values}
-              width={popupWidth}
-              emptyMessage={emptyMessage}
-              singleSelect={singleSelect}
-              onChangeValues={handleClickPanelItem}
-            />
-          </WizHStack>
-        </WizPopup>
-      )}
+      <WizPopup
+        anchorElement={inputRef}
+        isOpen={isPopupOpen}
+        onClose={() => setIsPopupOpen(false)}
+        isDirectionFixed={isDirectionFixed}
+      >
+        <WizHStack nowrap>
+          <SearchPopupPanel
+            options={filteredOptions}
+            closePopup={() => setIsPopupOpen(false)}
+            values={values}
+            width={popupWidth}
+            emptyMessage={emptyMessage}
+            singleSelect={singleSelect}
+            onChangeValues={handleClickPanelItem}
+          />
+        </WizHStack>
+      </WizPopup>
     </>
   );
 };