+++ packages/wiz-ui-next/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @wizleap-inc/wiz-ui-next
 
+## 2.43.0
+
+### Minor Changes
+
+- [#1377](https://github.com/Wizleap-Inc/wiz-ui/pull/1377) [`4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8`](https://github.com/Wizleap-Inc/wiz-ui/commit/4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(icon): filled counter
+
+### Patch Changes
+
+- Updated dependencies [[`4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8`](https://github.com/Wizleap-Inc/wiz-ui/commit/4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8)]:
+  - @wizleap-inc/wiz-ui-constants@0.31.0
+  - @wizleap-inc/wiz-ui-styles@0.33.1
+
+## 2.42.0
+
+### Minor Changes
+
+- [#1375](https://github.com/Wizleap-Inc/wiz-ui/pull/1375) [`ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e`](https://github.com/Wizleap-Inc/wiz-ui/commit/ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Chore(infomation-panel): update close button position
+
+- [#1374](https://github.com/Wizleap-Inc/wiz-ui/pull/1374) [`576c3c1b9795136d3df465a989bb4d8eef520b5b`](https://github.com/Wizleap-Inc/wiz-ui/commit/576c3c1b9795136d3df465a989bb4d8eef520b5b) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(icons): add counter, tunes, cloud_upload
+
+### Patch Changes
+
+- Updated dependencies [[`ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e`](https://github.com/Wizleap-Inc/wiz-ui/commit/ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e), [`576c3c1b9795136d3df465a989bb4d8eef520b5b`](https://github.com/Wizleap-Inc/wiz-ui/commit/576c3c1b9795136d3df465a989bb4d8eef520b5b)]:
+  - @wizleap-inc/wiz-ui-styles@0.33.0
+  - @wizleap-inc/wiz-ui-constants@0.30.0
+
+## 2.41.1
+
+### Patch Changes
+
+- [#1372](https://github.com/Wizleap-Inc/wiz-ui/pull/1372) [`c041edba88dbd007b3af58a541c30ba090d87898`](https://github.com/Wizleap-Inc/wiz-ui/commit/c041edba88dbd007b3af58a541c30ba090d87898) Thanks [@ichi-h](https://github.com/ichi-h)! - fix(wiz-ui-next): ジェネリクスの型がうまく生成できない問題の解消
+
+## 2.41.0
+
+### Minor Changes
+
+- [#1345](https://github.com/Wizleap-Inc/wiz-ui/pull/1345) [`73c5b87cfc3c75239de8814e97a26c42d1b3fcff`](https://github.com/Wizleap-Inc/wiz-ui/commit/73c5b87cfc3c75239de8814e97a26c42d1b3fcff) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(selects): ジェネリックコンポーネント化
+
 ## 2.40.0
 
 ### Minor Changes